backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
langchain-community
chromadb
duckduckgo-search
pydantic
